server/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==1.1.2
gunicorn==20.0.4
numpy
opencv-python
kafka-python
Flask-SocketIO==4.2.1
pykafka
stomp.py
